+++ internal/detect/tty_darwin.go
@@ -1,0 +1,7 @@
+//go:build darwin
+
+package detect
+
+import "syscall"
+
+const ioctlReadTermios = syscall.TIOCGETA