+++ internal/detect/tty_linux.go
@@ -1,0 +1,7 @@
+//go:build linux
+
+package detect
+
+import "syscall"
+
+const ioctlReadTermios = syscall.TCGETS